adding papers to conditions browse, various admin view fixes
