+++ yeastphenome/apps/conditions/templates/conditions/graphs/browse.html
@@ -2,10 +2,6 @@
 {% block title %}{{ object|capfirst }} / Experimental conditions / {{ SITE_NAME }}{% endblock %}
 {% block container %}
 <style>
-form {
-  font-family: sans-serif;
-}
-
 rect {
   cursor: pointer;
 }
@@ -20,6 +16,10 @@ svg {
     <div class="col-md-12">
         <h1>Browse Conditions</h1>
         <small>Top 100 By number of associated datasets</small>
+	<form style='padding-top:10px'>
+	  <label><input type="radio" name="mode" value="datasets" checked> Datasets</label>
+	  <label><input type="radio" name="mode" value="papers"> Papers</label>
+	</form>
     </div>
 </div>
 <div class="row" style="margin-bottom:200px">
@@ -30,8 +30,8 @@ svg {
 <script src="https://d3js.org/d3.v4.min.js"></script>
 <script>
 // {packageName: name, className: node.name, value: node.size});
-nodes = [{% for entry in data %}{% if entry.name != "standard" %}{"name": "{{ entry.name }}", "size": {{entry.number_of_datasets}}, "uid": {{ entry.id }} }{% if forloop.last %}{% else %},{% endif %}{% endif %}{% endfor %}]
-      
+nodes = [{% for entry in data %}{% if entry.name != "standard" %}{"name": "{{ entry.name }}", "size": {{entry.number_of_datasets}}, "uid": {{ entry.id }}, "papers": {{ entry.number_of_papers }} }{% if forloop.last %}{% else %},{% endif %}{% endif %}{% endfor %}]
+    
 var svg = d3.select("svg"),
     width = +svg.attr("width"),
     height = +svg.attr("height");
@@ -82,7 +82,10 @@ var treemap = d3.treemap()
       .text(function(d) { return d; });
 
   cell.append("title")
-      .text(function(d) { return d.data.id + "\n" + format(d.value); });
+      .text(function(d) { 
+         var choice = $("input[name='mode']:checked").val();
+         return d.data.id + "\n" + format(d.value) + " " + choice; 
+       });
 
   d3.selectAll("input")
       .data([sumBySize, sumByCount], function(d) { return d ? d.name : this.value; })
@@ -94,7 +97,27 @@ var treemap = d3.treemap()
         .dispatch("change");
   }, 2000);
 
+  $("input[name='mode']").change(function(){
+    choice = $(this).val()
+    console.log(choice)
+
+    if (choice=="datasets") {
+        treemap(root.sum(sumBySize));
+    } else {
+        treemap(root.sum(sumByPapers));
+    }
+
+    cell.transition()
+        .duration(750)
+        .attr("transform", function(d) { return "translate(" + d.x0 + "," + d.y0 + ")"; })
+      .select("rect")
+        .attr("width", function(d) { return d.x1 - d.x0; })
+        .attr("height", function(d) { return d.y1 - d.y0; });
+
+  })
+
   function changed(sum) {
+    console.log(sum);
     timeout.stop();
 
     treemap(root.sum(sum));
@@ -109,6 +132,10 @@ var treemap = d3.treemap()
 
 function sumByCount(d) {
   return d.children ? 0 : 1;
+}
+
+function sumByPapers(d) {
+  return d.papers;
 }
 
 function sumBySize(d) {
