+++ yeastphenome/apps/papers/forms.py
@@ -1,4 +1,4 @@
-from django.forms import ModelForm, Textarea
+from django.forms import ModelForm, Textarea, TextInput
 from .models import Paper
 
 
@@ -10,4 +10,6 @@ class PaperModelForm(ModelForm):
             "notes": Textarea(attrs={"cols": 85}),
             "private_notes": Textarea(attrs={"cols": 85}),
             "data_abstract": Textarea(attrs={"cols": 85, "rows": 3}),
+            "pmid": TextInput,
+            "pub_date": TextInput,
         }
